1_prompt_experiment/requirements.txt
openai==1.58.1
python-dotenv==1.0.1
editdistance==0.8.1
numpy==2.2.2